Add files via upload
This project is a Python-based implementation of the popular game "Rock, Paper, Scissors". It provides the opportunity to play the game against a computer, which makes its moves based on random choice. The game continues indefinitely until the user decides to quit. 
The game can be played by running the game.py script. The game continues until the user decides to quit by entering 'q'.